+++ database/migrations/2023_06_12_204229_create_member_base_table.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Models\Import\ImportedMember;
 use App\Models\Member;
 use App\Models\MemberGroup;
 use App\Models\User;