+++ database/migrations/2024_03_31_123054_sponsoring_adaptions_ad_placement.php
@@ -7,6 +7,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
+//TODO move into sponsoring migration after live deployment
 return new class extends Migration {
     /**
      * Run the migrations.