+++ database/migrations/2024_04_06_000606_add_member_info_column.php
@@ -1,0 +1,27 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+//TODO move into sponsoring migration after live deployment
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::table('members', function (Blueprint $table) {
+            $table->text('info')->nullable()->after('city');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropColumns('members', 'info');
+    }
+};